move unique payments to common types
