+++ account-abstraction/src/common/mod.rs
@@ -1,5 +1,4 @@
 pub mod common_types;
 pub mod custom_callbacks;
 pub mod signature;
-pub mod unique_payments;
 pub mod users;
